% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali.R
\name{cali}
\alias{cali}
\title{NMR spectral TSP calibration}
\usage{
cali(x, p)
}
\arguments{
\item{x}{The spectrum you want to calibrate as an array}

\item{p}{The matched ppm variable to the x you want to calibrate}
}
\value{
A spectrum that is calibrated to tsp
}
\description{
This function shifts the x varibale left or right so that it's TSP peak sits at 0 ppm.
}
\details{
\code{cali()} finds how far away the maximum value of of x between -0.2 and 0.2 is from the ppm variable closest to 0.
It then adds and subtracts the appropriate amount of values on either end of the spectrum so that x's TSP peak resides at 0.
}
\examples{
readin(path = system.file('extdata/15', package = 'NMRadjustr'))
xc <- cali(x, p)
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
