% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citra.R
\name{citra}
\alias{citra}
\title{Citrate Area Estimation}
\usage{
citra(x, p, n, sh = c(5.2, 5.3))
}
\arguments{
\item{x}{The spectrum of which you want to calculate the total area}

\item{p}{The matched ppm variable to x}

\item{n}{The noise estimation from the pproc}

\item{sh}{The ppm lower and upper limits defining the cirtate signal}
}
\value{
An estimated value of the citrate area
}
\description{
Estimates the area of the citrate signal
}
\details{
\code{citra()} finds the ppm values between the upper and lower bounds provided by the sh parameter.
Then it removes all values below the estimated noise level provided by \code{pproc()} to increase accuracy.
Finally the remaining values are summed and is the value returned by the function.
}
\examples{
data(x,p,n)
cit <- citra(x, p, n)
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
