% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea.R
\name{crea}
\alias{crea}
\title{Creatinine Area Estimator}
\usage{
crea(x, p, n, c3 = c(3, 3.1), c4 = c(4, 4.1))
}
\arguments{
\item{x}{The spectrum of which you want to calculate the total area}

\item{p}{The matched ppm variable to x}

\item{n}{The noise estimation from the pproc}

\item{c3}{The ppm lower and upper limits defining the creatinine signal at ppm 3}

\item{c4}{The ppm lower and upper limits defining the creatinine signal at ppm 4}
}
\value{
An estimated value for the creatinine area
}
\description{
Estimates the area of the creatinine signal
}
\details{
\code{crea()} finds the ppm values between the upper and lower bounds provided by the sh parameter.
Then it removes all values below the estimated noise level provided by \code{pproc()} to increase accuracy.
Finally the remaining values are summed and is the value returned by the function.
}
\examples{
data(x,p,n)
cr <- crea(x, p, n)
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
