% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_.R
\name{flip_}
\alias{flip_}
\title{Spectral Flipping}
\usage{
flip_(x, p, sh)
}
\arguments{
\item{x}{the spectra wanting to be flipped}

\item{p}{the match ppm variable}

\item{sh}{the chemical shift used to calculate orientation given as the upper and lower ppm bounds}
}
\value{
Returns a spectrum (array) with the correct orientation
}
\description{
\code{flip_()} re-orientates/flips the spectrum on the x axis so that the peaks are pointing the correct way.
}
\details{
\code{flip_()} looks at the chemical shift provided by the sh variable and sums the x values in that area.
It then determines if the sum of this area has a positive or negative sign (indicating if in that area, the peak is pointing in the positive or negative direction). If it is in the negative direction, the entire spectrum is mulitplied by -1 and returned.
}
\examples{
readin(path = system.file('extdata/15', package = 'NMRadjustr'))
xf <- flip_(x, p, sh = c(3,3.1))
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
