% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\docType{data}
\name{p}
\alias{p}
\title{The ppm of the NMR spectrum.}
\format{
an array with 55674 elements with their value pertaining to the ppm frequency
}
\usage{
p
}
\description{
An array of values containing information about frequency of a metabolite signal.
}
\keyword{datasets}
