% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproc.R
\name{pproc}
\alias{pproc}
\title{Spectral Preprocessing}
\usage{
pproc(x, p)
}
\arguments{
\item{x}{The spectra that you are processing}

\item{p}{The ppm matched to your spectra}
}
\value{
This function returns:
\enumerate{
\item \strong{x}: a fully preprocessed spectrum ready for it's dilution to be estimated
\item \strong{p}: the matched ppm variable for the preprocessed x
\item \strong{o}: the original X spectra that was input into the function
\item \strong{n}: the estimated value of noise for this spectrum
}
}
\description{
\code{pproc()} streamlines the processing of NMR spectra by combining functions of NMRadjustr in the appropriate order to return the cleanest spectrum possible with the least amount of fuss.
}
\details{
\code{pproc()} only takes one spectrum at a time due to the nature of NMRadjustr's purpose.
It begins be ensuring the spectrum is orientated the right way. Often spectra acquired with a small number of scans are incorrectly orientated because the water signal towers over all other signals.
From there, the spectra are calibrated so that the peak of the TSP signal is at a ppm of 0.
The lower ppm, water and urea regions are then removed, leaving the upper ppm region for a noise estimation.
Baseline correction using asymmetric least squares is performed and a noise estimation between the region of 9.5/11 is done before the upper ppm region is removed.
Finally, the processed x spectrum, matched ppm variable, original spectrum and noise estimation are then assigned to the global environment overwriting the existing varibales.
}
\examples{
readin(path = system.file('extdata/15', package = 'NMRadjustr'))
pproc(x, p)
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
