% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin}
\alias{readin}
\title{Import 1D NMR spectra}
\usage{
readin(path)
}
\arguments{
\item{path}{Given as a string, the path to the overarching files containing the NMR spectrum.}
}
\value{
The function exports the following three objects into the currently active R environment (no variable assignments needed):
\enumerate{
\item \strong{x}: The NMR spectrum in an array of values matched to p
\item \strong{p}: The column-matched ppm array of the x variable
\item \strong{m}: The spectrometer metadata as extracted from the \emph{acqus} file, row-matched to x
}
}
\description{
Import 1D NMR spectra
}
\details{
This function imports TopSpin processed NMR spectra as well as spectrometer and processing parameters found in files \emph{acqus} and \emph{procs}. Experiments can be filtered according to data acquisition variables using the \code{exp_type} argument: For example, to read standard 1D NMR experiments use \code{exp_type=list(exp='noesygppr1d')}. More than one argument can be provided as list element. \strong{Objects in the R environment with the same variable names will be overwritten.}
}
\section{}{
NA
}

\examples{
readin(path = system.file('extdata/15', package = 'NMRadjustr'))
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
