% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundr.R
\name{roundr}
\alias{roundr}
\title{Round to a Multiple of 4}
\usage{
roundr(n, r2 = c(4, 512, 4))
}
\arguments{
\item{n}{The number you want to round}

\item{r2}{Encodes the values plugged into the \code{seq()} function with the elements being:
\enumerate{
\item The minimum value to be rounded to
\item The maximum value to be rounded to, and
\item The step/multiple to be incremented by. Ideally, 1 and 2 should be a multiple of 3
}}
}
\value{
A integer rounded to the nearest multiple of the third element in r2
}
\description{
This function rounds the given value n to the nearest value created by r2.
}
\details{
\code{roundr()} works by creating an array and finding which number n is closest to.
r2 contains all the input arguments for the \code{seq()} function (being the min, max and step args) which are then parsed to \code{seq()} to create an array of values.
\code{roundr()} then takes the number to be round: n, and finds the minimum absolute difference between it and all values in the sequence created by r2.
Finally the number in the array closest to n is returned as the 'rounded' value.
}
\examples{
ns <- roundr(n = 37, r2 = c(4,256,4))
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
