% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\docType{data}
\name{x}
\alias{x}
\title{Processed NMR urine spectrum.}
\format{
an array with 55674 elements with their value pertaining to the ppm intensity
}
\usage{
x
}
\description{
An array of values containing information about metabolites within a urine sample.
}
\keyword{datasets}
